requirements.txt
django
djangorestframework
django-cors-headers
djoser
pillow
gunicorn
openai
python-decouple
drf-spectacular